Add last calculated cache size to Dashboard Storage chart (#1580)
issue #1580 Add last calculated cache size to Dashboard Storage chart

Size of cached files with all derivatives and representatives may be
as large as all image files, actually doubling the needed storage
size. So it may be nice to see it represented in the storage
chart, if one remembers to occasionally recalculate it under
Tools -> Maintenance..

Also make it possible to deactivate this setting.