issue #2073 add raw_name in tags property
We present the raw version in the input, and on the rendered version there's a small globe to show that there's a difference between the rendered and raw versions.

We have also modified the album title in the album edit page. The rendered version is now displayed, we keep the raw version in the input field.