fixes #2202 smarter filters for gallery search
To be able to compute values of filter "added_by" by only taking into account
results of other filters, we need to distinct results filter by filter. So
instead of generating a single complex SQL query, we use one (or maybe more)
SQL query for each filter in action.

First commit for this new way to perform a regular search. To be improved
(many duplication of codes for now)